Added a new method to IGlastDetSvc and GlastDetSvc to wrap a new IDmapBuilder method to retrieve shape dimensions and type from the id of a volume; the new method is called getShapeByID.
Fixed a small bug in the DMManager init method for the setting of the mode

Removed a sign reversal for rotation from the GeometryVisitor. In this way rotation from
GlastDetSvc have the same convention (sign) of rotation in xmlGeoDbs. It is up to the clients
(like G4Generator) to change this sign if conventions are different.